requirements.txt
Django
django-cors-headers
djangorestframework
instaloader
joblib
matplotlib
numpy
pandas
scikit-learn
seaborn
requests
Pillow
opencv-python-headless
ImageHash
openai-whisper
ffmpeg-python
gunicorn
whitenoise